pyproject.toml
[project]
# Only needed so uv/pip recognizes it as a project
name = "ex-agents-patterns"
version = "0.0.1"
description = "LangGraph Essentials - Python"
requires-python = ">=3.11,<3.14"

# Dependencies you need in the notebooks
dependencies = [
    "langgraph>=1.0.2",
    "langchain>=1.0.3",
    "langchain-openai>=1.0.1",
    "langchain-anthropic>=1.0.0",
    "langchain-community",
    "jupyter>=1.0.0",
    "langgraph-cli[inmem]>=0.4.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
# Extra tools for development
dev = [
    "ruff>=0.6.1",
    "mypy>=1.11.1"
]

[tool.ruff]
# What to check
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # docstring first line imperative
    "UP",   # pyupgrade
]

# Rules to ignore
lint.ignore = [
    "T201", # allow print()
    "UP006", # keep using typing.List etc.
    "UP007", # allow Union[X, Y]
    "UP035", # ignore Self type suggestion
    "D417",  # don't require param docs
    "E501",  # ignore line length
    "D101",  # ignore docstring for these labs
    "D103",  # ignore docstring for these labs
    "D202",  # ignore extra space after docstring requirement
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"

